Potential fix for code scanning alert no. 1: Multiplication result converted to larger type

+++ src/libdstdec/dst_init.c
@@ -87,7 +87,7 @@ static void *MemoryAllocate(int NrOfElements, int SizeOfElement)
     fprintf(stderr,"ERROR: not enough memory available!\n\n");
   }
   #else
-  if ((Array = _mm_malloc(NrOfElements * SizeOfElement, 16)) == NULL) 
+  if ((Array = _mm_malloc((size_t)NrOfElements * SizeOfElement, 16)) == NULL) 
   {
     fprintf(stderr,"ERROR: not enough memory available!\n\n");
   }
